+++ src/components/NavBar/index.tsx
@@ -1,24 +1,19 @@
-import { Trans } from '@lingui/macro'
-import { useWeb3React } from '@web3-react/core'
+/* eslint-disable react/prop-types */
 import Web3Status from 'components/Web3Status'
-import { chainIdToBackendName } from 'graphql/data/util'
 import { useIsNftPage } from 'hooks/useIsNftPage'
-import { useIsPoolsPage } from 'hooks/useIsPoolsPage'
-import { useAtomValue } from 'jotai/utils'
 import { Box } from 'nft/components/Box'
 import { Row } from 'nft/components/Flex'
 import { UniIcon } from 'nft/components/icons'
 import { useProfilePageState } from 'nft/hooks'
 import { ProfilePageStateType } from 'nft/types'
 import { ReactNode } from 'react'
-import { NavLinkProps, useLocation, useNavigate } from 'react-router-dom'
-import { shouldDisableNFTRoutesAtom } from 'state/application/atoms'
+import { NavLinkProps, useNavigate } from 'react-router-dom'
 import styled from 'styled-components/macro'
 
 import { Bag } from './Bag'
 import Blur from './Blur'
 import { ChainSelector } from './ChainSelector'
-import { MenuDropdown } from './MenuDropdown'
+import { PageTabs } from './PageTabs'
 import { SearchBar } from './SearchBar'
 import * as styles from './style.css'
 
@@ -42,65 +37,35 @@ interface MenuItemProps {
   dataTestId?: string
   id?: string
   isActive?: boolean
+  isTarget?: boolean
   children: React.ReactNode
 }
 
-const MenuItem: React.FC<MenuItemProps> = ({ href, dataTestId, id, isActive, children }) => {
-  const isExternalLink = href.startsWith('https') || href.startsWith('//')
+export const MenuItem: React.FC<MenuItemProps> = ({ href, dataTestId, id, isActive, children, isTarget = true }) => {
+  // const isExternalLink = href.startsWith('https') || href.startsWith('//')
 
-  const linkProps = isExternalLink
-    ? {
-        href,
-        target: '_blank',
-        rel: 'noopener noreferrer',
-      }
-    : {
-        to: href,
-      }
+  // const linkProps = isExternalLink
+  //   ? {
+  //       href,
+  //       target: '_blank',
+  //       rel: 'noopener noreferrer',
+  //     }
+  //   : {
+  //       to: href,
+  //     }
 
   return (
     <a
-      {...linkProps}
+      href={href}
+      target={isTarget == true ? '_blank' : ''}
       className={isActive ? styles.activeMenuItem : styles.menuItem}
       id={id}
       style={{ textDecoration: 'none' }}
       data-testid={dataTestId}
+      rel="noreferrer"
     >
       {children}
     </a>
-  )
-}
-
-export const PageTabs = () => {
-  const { pathname } = useLocation()
-  const { chainId: connectedChainId } = useWeb3React()
-  const chainName = chainIdToBackendName(connectedChainId)
-
-  const isPoolActive = useIsPoolsPage()
-  const isNftPage = useIsNftPage()
-
-  const shouldDisableNFTRoutes = useAtomValue(shouldDisableNFTRoutesAtom)
-
-  return (
-    <>
-      <MenuItem href="https://3000-alfset-physicainterfae-z1ncymx2xii.ws-us106.gitpod.io/#/swap">
-        <Trans>Swap</Trans>
-      </MenuItem>
-      <MenuItem href="https://github.com/alfset/Portal-Bridge">
-        <Trans>Stake</Trans>
-      </MenuItem>
-      <MenuItem href="https://github.com/alfset/Portal-Bridge">
-        <Trans>Bridge</Trans>
-      </MenuItem>
-      <Box display={{ sm: 'flex', lg: 'none', xxl: 'flex' }} width="full">
-        <MenuItem href="https://5173-deltaswapio-deltaswap-vl9t8cfk52f.ws-us106.gitpod.io/">
-          <Trans>Pools</Trans>
-        </MenuItem>
-      </Box>
-      <Box marginY={{ sm: '4', md: 'unset' }}>
-        <MenuDropdown />
-      </Box>
-    </>
   )
 }
